deploy/requirements.txt
line-bot-sdk==3.14.2
openai==1.57.4
google-cloud-firestore==2.19.0
firebase_admin==6.6.0